test: migrate some driver tests to the new testing
